# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Battle Rap API
  version: 1.0.0
  description: >
    REST API for managing online rap battles, tournaments, submissions, judging,
    and administration workflows.
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/v1/challenges:
    get:
      tags: [Challenges]
      summary: List challenges
      operationId: listChallenges
      responses:
        '200':
          description: Active and recent challenges.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Challenge'
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags: [Challenges]
      summary: Create challenge
      operationId: createChallenge
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [opponent_id, title]
              properties:
                opponent_id:
                  type: string
                  format: uuid
                title:
                  type: string
                  minLength: 3
                description:
                  type: string
                  nullable: true
      responses:
        '201':
          description: Challenge created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Challenge'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token issued by the API. Required for authenticated endpoints.
    RefreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: >
        Refresh token cookie issued during login/registration.
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    Problem:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        detail:
          type: string
        status:
          type: integer
        field_errors:
          type: object
          additionalProperties:
            type: string
      example:
        type: https://httpstatuses.com/422
        title: Payload validation failed.
        detail: Payload validation failed.
        status: 422
        field_errors:
          email: Invalid email.
...
    ChallengeUserRef:
      type: object
      required:
        - id
        - display_name
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
    ChallengeVotes:
      type: object
      required:
        - initiator
        - opponent
      properties:
        initiator:
          type: integer
          minimum: 0
        opponent:
          type: integer
          minimum: 0
    Challenge:
      type: object
      required:
        - id
        - title
        - status
        - initiator
        - opponent
        - votes
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [initiated, in_progress, completed, cancelled]
        timestamps:
          type: object
          properties:
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            accepted_at:
              type: string
              format: date-time
              nullable: true
            started_at:
              type: string
              format: date-time
              nullable: true
            completed_at:
              type: string
              format: date-time
              nullable: true
            cancelled_at:
              type: string
              format: date-time
              nullable: true
        initiator:
          $ref: '#/components/schemas/ChallengeUserRef'
        opponent:
          $ref: '#/components/schemas/ChallengeUserRef'
        votes:
          $ref: '#/components/schemas/ChallengeVotes'
...